+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 ## 4.0.0 (UNRELEASED)
 ### 📝 Notes
+* The support for Nextcloud versions below 30 has been removed,
+  this means some functions like filename validation will now only
+  work with the capabilities provided by Nextcloud 30 or newer.
+
 #### DAV related export
 The DAV related exports from the main entry point were deprecated
 for a long time and are now removed from it.